+++ src/psytricks/cli.py
@@ -129,6 +129,7 @@ def run_cli(
     command : str
         The command indicating which wrapper method to call.
     """
+    # FIXME: adapt for "kwargs" being dropped in PSyTricksWrapper
     configure_logging(verbose)
     wrapper = PSyTricksWrapper(deliverycontroller=cdc)
     call_method = {